完善ui
1.加入插件编辑界面（未完成）
2.加入插件的资源元素、库元素（未完成）
3.缩减变量名

+++ DesktopPony/data/text.cpp
@@ -55,7 +55,7 @@ QString Text::getPara(QString para, QVariant var)
     } else if(key == QStringLiteral("var")) {
         qint32 s = value.toInt();
         s = s < 1 ? 1 : s;
-        return varToString(getVar(var, s).toInt());
+        return varToString(getVar(var, s));
     } else {
         return nullptr;
     }
